Add useEditController
